requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow
matplotlib
opencv-python
numpy
tqdm
tensorboard
ptflops
timm
torchvision